requirements.txt
requests==2.31.0
influxdb-client==1.36.0
python-dotenv==1.0.0
pyyaml==6.0
requests>=2.28.0
websocket-client>=1.6.0
python-dotenv>=1.0.0
